refactor(backend): preparation to move exporter to different service
