+++ apps/backend/src/importer/generic_json.rs
@@ -3,12 +3,9 @@ use std::{fs, sync::Arc};
 use async_graphql::Result;
 use enums::ImportSource;
 use itertools::Itertools;
-use models::CompleteExport;
+use models::{CompleteExport, DeployJsonImportInput};
 
-use crate::{
-    fitness::resolver::ExerciseService,
-    importer::{DeployJsonImportInput, ImportResult},
-};
+use crate::{fitness::resolver::ExerciseService, importer::ImportResult};
 
 pub async fn import(
     input: DeployJsonImportInput,
