+++ apps/backend/src/importer/goodreads.rs
@@ -5,17 +5,15 @@ use csv::Reader;
 use enums::{ImportSource, MediaLot, MediaSource};
 use itertools::Itertools;
 use models::{
-    ImportOrExportItemRating, ImportOrExportItemReview, ImportOrExportMediaItem,
-    ImportOrExportMediaItemSeen,
+    DeployGenericCsvImportInput, ImportOrExportItemRating, ImportOrExportItemReview,
+    ImportOrExportMediaItem, ImportOrExportMediaItemSeen,
 };
 use providers::google_books::GoogleBooksService;
 use rust_decimal::Decimal;
 use rust_decimal_macros::dec;
 use serde::Deserialize;
 
-use crate::importer::{
-    DeployGenericCsvImportInput, ImportFailStep, ImportFailedItem, ImportResult,
-};
+use crate::importer::{ImportFailStep, ImportFailedItem, ImportResult};
 
 #[derive(Debug, Deserialize)]
 struct Book {
