+++ apps/backend/src/importer/imdb.rs
@@ -2,13 +2,11 @@ use async_graphql::Result;
 use csv::Reader;
 use enums::{MediaLot, MediaSource};
 use itertools::Itertools;
-use models::{DefaultCollection, ImportOrExportMediaItem};
+use models::{DefaultCollection, DeployGenericCsvImportInput, ImportOrExportMediaItem};
 pub use providers::tmdb::NonMediaTmdbService;
 use serde::Deserialize;
 
-use crate::importer::{
-    DeployGenericCsvImportInput, ImportFailStep, ImportFailedItem, ImportResult,
-};
+use crate::importer::{ImportFailStep, ImportFailedItem, ImportResult};
 
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "PascalCase")]
