+++ apps/backend/src/importer/jellyfin.rs
@@ -1,7 +1,9 @@
 use async_graphql::Result;
 use enum_meta::HashMap;
 use enums::{MediaLot, MediaSource};
-use models::{ImportOrExportMediaItem, ImportOrExportMediaItemSeen};
+use models::{
+    DeployUrlAndKeyAndUsernameImportInput, ImportOrExportMediaItem, ImportOrExportMediaItemSeen,
+};
 use reqwest::{
     header::{HeaderMap, HeaderValue, ACCEPT, AUTHORIZATION, USER_AGENT},
     Client, ClientBuilder,
@@ -11,9 +13,7 @@ use serde::{Deserialize, Serialize};
 use serde_json::json;
 use utils::{APPLICATION_JSON_HEADER, USER_AGENT_STR};
 
-use crate::importer::{
-    DeployUrlAndKeyAndUsernameImportInput, ImportFailStep, ImportFailedItem, ImportResult,
-};
+use crate::importer::{ImportFailStep, ImportFailedItem, ImportResult};
 
 static EMBY_HEADER_VALUE: &str =
     r#"MediaBrowser , Client="other", Device="script", DeviceId="script", Version="0.0.0""#;
