+++ apps/backend/src/importer/media_tracker.rs
@@ -1,8 +1,8 @@
 use async_graphql::Result;
 use enums::{ImportSource, MediaLot, MediaSource};
 use models::{
-    CreateOrUpdateCollectionInput, IdObject, ImportOrExportItemRating, ImportOrExportItemReview,
-    ImportOrExportMediaItemSeen,
+    CreateOrUpdateCollectionInput, DeployUrlAndKeyImportInput, IdObject, ImportOrExportItemRating,
+    ImportOrExportItemReview, ImportOrExportMediaItemSeen,
 };
 use providers::openlibrary::get_key;
 use reqwest::{
@@ -16,10 +16,7 @@ use serde::{Deserialize, Serialize};
 use serde_with::{formats::Flexible, serde_as, TimestampMilliSeconds};
 use utils::USER_AGENT_STR;
 
-use crate::importer::{
-    DeployUrlAndKeyImportInput, ImportFailStep, ImportFailedItem, ImportOrExportMediaItem,
-    ImportResult,
-};
+use crate::importer::{ImportFailStep, ImportFailedItem, ImportOrExportMediaItem, ImportResult};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 #[serde(rename_all = "snake_case")]
