+++ crates/background/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "background"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+apalis = { workspace = true }
+common_models = { path = "../models/common" }
+enums = { path = "../enums" }
+fitness_models = { path = "../models/fitness" }
+media_models = { path = "../models/media" }
+serde = { workspace = true }
+strum = { workspace = true }
+uuid = { workspace = true }
